ai-service/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
supabase
opencv-python
numpy
websockets
python-multipart
vision-agents[moondream, gemini]
Pillow
